requirements.txt
opencv-python
numpy
Pillow
google-generativeai